pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "proteinconformers"
version = "0.1.0"
description = "Benchmark tooling for simulating and evaluating protein conformational landscapes."
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = { file = "LICENSE" }
authors = [
    { name = "ProteinConformers Team" }
]
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "tqdm>=4.66",
    "torch>=2.1",
    "torchmetrics>=1.2",
    "einops>=0.7",
    "transformers>=4.37",
    "tokenizers>=0.15",
    "huggingface-hub>=0.20",
    "lightning>=2.1",
    "omegaconf>=2.3",
    "hydra-core>=1.3",
    "MDAnalysis>=2.6",
    "mdtraj>=1.9",
    "biopython>=1.83",
    "biotite>=0.38",
    "cloudpathlib>=0.16",
    "brotli>=1.1",
    "msgpack>=1.0",
    "msgpack-numpy>=0.4",
    "deeptime>=0.4",
    "esm==3.0.4",
    "dm-tree>=0.1",
    "zstd>=1.5"
]

[project.optional-dependencies]
bioemu = ["bioemu>=0.1"]

[project.urls]
Repository = "https://github.com/ProteinConformers/ProteinConformers"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4",
    "ruff>=0.3"
]
